# contracts/openapi/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Evidrilo API",
    "version": "1.0.0",
    "description": "Public contract for the versioned Evidrilo API routes."
  },
  "paths": {
...
    "/v1/project-ai/stage-assist/settlement": {
      "post": {
        "operationId": "post_v1_project_ai_stage_assist_settlement",
        "tags": [
          "project-ai"
        ],
        "security": [
          {
            "SupabaseBearer": []
          }
        ],
        "responses": {
          "2XX": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/project-ai-stage-assist-settlement.v1"
                }
              }
            }
          },
          "default": {
            "description": "API error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/http-errors.v1"
                }
              }
            }
          }
        },
        "x-evidrilo-test": {
          "file": "platform/api.Tests/ProjectAiScaffoldFlowTests.cs",
          "anchor": "/v1/project-ai/stage-assist/settlement"
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/project-ai-stage-assist-settlement-request.v1"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Stable request key used to make this operation idempotent.",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9_-]+$"
            }
          }
        ]
      },
      "parameters": []
    },
...
  },
  "components": {
    "securitySchemes": {
      "SupabaseBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "RevenueCatWebhookAuthorization": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      },
      "RevenueCatWebhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RevenueCat-Webhook-Signature"
      }
    },
    "schemas": {
...
      "http-errors.v1": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "https://evidrilo.dev/contracts/http-errors.v1.json",
        "title": "Evidrilo HTTP error response v1",
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema",
          "version",
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "schema": {
            "const": "evidrilo.http-error"
          },
          "version": {
            "const": "1"
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]{2,63}$"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 240
          },
          "requestId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "pattern": "^[A-Za-z0-9_-]+$"
          },
          "creditCost": {
            "description": "Present only when verified provider usage was charged even though the requested response was withheld.",
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          }
        }
      },
...
      "project-ai-stage-assist-settlement-request.v1": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "https://evidrilo.dev/contracts/project-ai-stage-assist-settlement-request.v1.json",
        "title": "Evidrilo stage-assist settlement request v1",
        "description": "The Idempotency-Key header must match requestId from the preview. Reuse the identical body and key for retries. Applied or edited local-project outcomes must name the resulting revision and metadata binding generation.",
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema",
          "version",
          "installationId",
          "requestId",
          "outcome"
        ],
        "properties": {
          "schema": {
            "const": "evidrilo.project-ai-stage-assist-settlement"
          },
          "version": {
            "const": "1"
          },
          "installationId": {
            "type": "string",
            "format": "uuid"
          },
          "requestId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{8,128}$"
          },
          "outcome": {
            "enum": [
              "APPLIED",
              "EDITED",
              "DISMISSED",
              "STALE"
            ]
          },
          "resultProjectRevision": {
            "type": "integer",
            "minimum": 2
          },
          "resultProjectBindingGeneration": {
            "type": "integer",
            "minimum": 1
          }
        },
        "allOf": [
          {
            "if": {
              "properties": {
                "outcome": {
                  "enum": [
                    "APPLIED",
                    "EDITED"
                  ]
                }
              },
              "required": [
                "outcome"
              ]
            },
            "then": {
              "required": [
                "resultProjectRevision",
                "resultProjectBindingGeneration"
              ]
            }
          },
          {
            "if": {
              "properties": {
                "outcome": {
                  "enum": [
                    "DISMISSED",
                    "STALE"
                  ]
                }
              },
              "required": [
                "outcome"
              ]
            },
            "then": {
              "not": {
                "anyOf": [
                  {
                    "required": [
                      "resultProjectRevision"
                    ]
                  },
                  {
                    "required": [
                      "resultProjectBindingGeneration"
                    ]
                  }
                ]
              }
            }
          }
        ]
      },
      "project-ai-stage-assist-settlement.v1": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "https://evidrilo.dev/contracts/project-ai-stage-assist-settlement.v1.json",
        "title": "Evidrilo stage-assist settlement response v1",
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema",
          "version",
          "outcome",
          "requestId",
          "creditCost"
        ],
        "properties": {
          "schema": {
            "const": "evidrilo.project-ai-stage-assist-settlement"
          },
          "version": {
            "const": "1"
          },
          "outcome": {
            "enum": [
              "APPLIED",
              "EDITED",
              "DISMISSED",
              "STALE"
            ]
          },
          "requestId": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]{8,128}$"
          },
          "creditCost": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          }
        }
      },
...
    }
  }
}